patched wire library for avoid multimaster conflict on i2c bus

+++ arduino/sketchbook/hardware/Microduino/avr/libraries/Wire/src/Wire.h
@@ -72,7 +72,7 @@ class TwoWire : public Stream
     void onReceive( void (*)(int) );
     void onRequest( void (*)(void) );
 
-    inline size_t write(unsigned long n) { return write((uint8_t)n); }
+    inline size_t write(uint32_t n) { return write((uint8_t)n); }
     inline size_t write(long n) { return write((uint8_t)n); }
     inline size_t write(unsigned int n) { return write((uint8_t)n); }
     inline size_t write(int n) { return write((uint8_t)n); }
